[박소정] questionBank에서 문제 삭제할 때 exam_question에 데이터가 있을 경우, 외래키로 인해서 삭제가 되지 않는 오류가 발생했음 --> 삭제 서비스에 exam_question 테이블에 해당 문제가 들어가있는지 확인하는 mapper를 추가하고 예외 처리해서 오류 수정함/ questionBank.jsp에 알람메시지 오류500번 추가/ 단답형 수정 및 삭제 버튼 수정함

+++ lms/src/main/java/com/gd/lms/service/MultiplechoiceService.java
@@ -10,8 +10,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.gd.lms.commons.TeamColor;
+import com.gd.lms.mapper.ExamQuestionMapper;
 import com.gd.lms.mapper.MultiplechoiceExampleMapper;
 import com.gd.lms.mapper.MultiplechoiceMapper;
+import com.gd.lms.vo.ExamQuestion;
 import com.gd.lms.vo.Multiplechoice;
 import com.gd.lms.vo.MultiplechoiceExample;
 
@@ -28,6 +30,10 @@ public class MultiplechoiceService {
 	// MultiplechoiceExampleMapper 객체 주입
 	@Autowired
 	private MultiplechoiceExampleMapper multiplechoiceExampleMapper;
+
+	// ExamQuestionMapper 객체 주입
+	@Autowired
+	private ExamQuestionMapper examQuestionMapper;
 
 	// 객관식 문제 지문 수정하기
 	// 파라미터 : Multiplechoice
@@ -54,14 +60,25 @@ public class MultiplechoiceService {
 		// 파라미터 디버깅
 		log.debug(TeamColor.PSJ + questionNo + "<-- questionNo" + TeamColor.TEXT_RESET);
 
-		// 리턴 변수 선언 
+		// 리턴 변수 선언
 		int removeMultiplechoiceCk = 0;
 
-		// 1. multiplechoiceExample에서 보기 삭제 
-		int removeMultiplechoiceExampleCk = multiplechoiceExampleMapper.deleteMultiplechoiceExampleByQuestionNo(questionNo);
+		// exam_question에 multiplechoice 데이터가 포함되어 있으면 FK로 연결되어 있기 때문에 삭제가 안됨 --> exam_question테이블에 데이터가 들어가져있는 확인해야 한다
+		int CountCk = examQuestionMapper.selectCountExamQuestion(questionNo);
+		// 결과값 디버깅
+		if (CountCk != 0) {
+			// 디버깅
+			log.debug(TeamColor.PSJ + "[Error] exam_question 테이블에 데이터가 있어서 삭제가 불가능합니다. " + TeamColor.TEXT_RESET);
+			// 500으로 리턴해서 삭제 실패로 보내기
+			return 500;
+		}
+
+		// 1. multiplechoiceExample에서 보기 삭제
+		int removeMultiplechoiceExampleCk = multiplechoiceExampleMapper
+				.deleteMultiplechoiceExampleByQuestionNo(questionNo);
 		// 2. multiplechoice에서 문제 삭제
 		if (removeMultiplechoiceExampleCk != 0) { // 보기 삭제에 성공했다면
-			//디버깅
+			// 디버깅
 			log.debug(TeamColor.PSJ + "[Success] 해당 객관식 문제 보기 삭제 성공" + TeamColor.TEXT_RESET);
 			// 해당 객관식 문제 삭제
 			removeMultiplechoiceCk = multiplechoiceMapper.deleteMultiplechoiceByQuestionNo(questionNo);
@@ -73,7 +90,7 @@ public class MultiplechoiceService {
 				return 0;
 			}
 		} else { // 객관식 보기 삭제에 실패했다면
-			//디버깅
+			// 디버깅
 			log.debug(TeamColor.PSJ + "[Fail] 해당 객관식 문제 보기 삭제 실패" + TeamColor.TEXT_RESET);
 			return 0;
 		}
@@ -84,7 +101,7 @@ public class MultiplechoiceService {
 	// 파라미터 : questionNo
 	// 리턴값 : Map<String, Object>
 	public Map<String, Object> getMultiplechoiceOne(int questionNo) {
-		//파라미터 디버깅
+		// 파라미터 디버깅
 		log.debug(TeamColor.PSJ + questionNo + "<-- questionNo" + TeamColor.TEXT_RESET);
 
 		// 리턴값을 담을 객체 생성
@@ -94,7 +111,8 @@ public class MultiplechoiceService {
 		log.debug(TeamColor.PSJ + "객관식 문제 정보 가져오기 성공" + TeamColor.TEXT_RESET);
 
 		// MultiplechoiceExampleMapper에서 해당 객관식 문제의 보기 지문 받아오기
-		multiplechoiceOne.put("multiplechoiceExample", multiplechoiceExampleMapper.selectMultiplechoiceExampleList(questionNo));
+		multiplechoiceOne.put("multiplechoiceExample",
+				multiplechoiceExampleMapper.selectMultiplechoiceExampleList(questionNo));
 		log.debug(TeamColor.PSJ + "해당 객관식 문제의 보기 가져오기 성공" + TeamColor.TEXT_RESET);
 
 		return multiplechoiceOne;
@@ -102,7 +120,7 @@ public class MultiplechoiceService {
 	}
 
 	// 검색어와 관련된 객관식 문제 리스트 가져오기
-	// 파라미터 : subjectName 
+	// 파라미터 : subjectName
 	// 리턴값 : List<Multiplechoice>
 	public List<Map<String, Object>> getMultipleChoiceList(String subjectName) {
 		// 파라미터 디버깅
@@ -140,11 +158,11 @@ public class MultiplechoiceService {
 
 		// 전송받은 보기 지문을 list 생성 후 넣기
 		List<String> paramExampleList = new ArrayList<>();
-		paramExampleList.add((String) paramMap.get("answer1")); //인덱스 0
-		paramExampleList.add((String) paramMap.get("answer2")); //인덱스 1
-		paramExampleList.add((String) paramMap.get("answer3")); //인덱스 2
-		paramExampleList.add((String) paramMap.get("answer4")); //인덱스 3
-		paramExampleList.add((String) paramMap.get("answer5")); //인덱스 4
+		paramExampleList.add((String) paramMap.get("answer1")); // 인덱스 0
+		paramExampleList.add((String) paramMap.get("answer2")); // 인덱스 1
+		paramExampleList.add((String) paramMap.get("answer3")); // 인덱스 2
+		paramExampleList.add((String) paramMap.get("answer4")); // 인덱스 3
+		paramExampleList.add((String) paramMap.get("answer5")); // 인덱스 4
 
 		// MultiplechoiceExampleMapper에 전송할 파라미터를 List로 셋팅
 		List<MultiplechoiceExample> paramMultiplechoiceExampleList = new ArrayList<>();
